LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/text-combine-upright-parsing-valid-001-expected.txt


FAIL Computed value for `text-combine-upright: none` is `none` assert_equals: expected (string) "none" but got (undefined) undefined
FAIL Computed value for `text-combine-upright: all` is `all` assert_equals: expected (string) "all" but got (undefined) undefined
FAIL Computed value for `text-combine-upright: initial` is `none` assert_equals: expected (string) "none" but got (undefined) undefined
FAIL Computed value for `text-combine-upright: inherit` is `all` which is inherited from the parent element assert_equals: expected (string) "all" but got (undefined) undefined
FAIL Computed value for `text-combine-upright: unset` is `all` which is inherited from the parent element assert_equals: expected (string) "all" but got (undefined) undefined
